parser: Split experimental parser in mods
